refactor: move Cover into SharedComponents
