+++ src/components/MobileFilters/MobileFilters.tsx
@@ -13,7 +13,7 @@ import Button from "@/components/SharedComponents/Button";
 import ButtonAsLink from "@/components/SharedComponents/Button/ButtonAsLink";
 import Flex, { FlexProps } from "@/components/SharedComponents/Flex";
 import useEventListener from "@/hooks/useEventListener";
-import Cover from "@/components/Cover";
+import Cover from "@/components/SharedComponents/Cover";
 import { IconName } from "@/components/SharedComponents/Icon";
 import { useMenuContext } from "@/context/Menu";
 import { PostCategoryPage } from "@/components/Posts/PostCategoryList/PostCategoryList";
