+++ src/components/SharedComponents/Card/SummaryCard.tsx
@@ -2,12 +2,12 @@ import { FC } from "react";
 
 import { PortableTextJSON, Image } from "../../../common-lib/cms-types";
 import { OakColorName } from "../../../styles/theme/types";
-import Cover from "../../Cover";
 import BrushBorders from "../../SpriteSheet/BrushSvgs/BrushBorders";
 import { PortableTextWithDefaults } from "../../PortableText";
 
 import Card from "./Card";
 
+import Cover from "@/components/SharedComponents/Cover";
 import Typography, { Heading } from "@/components/SharedComponents/Typography";
 import CMSImage from "@/components/SharedComponents/CMSImage";
 import Flex, { FlexProps } from "@/components/SharedComponents/Flex";
