+++ src/components/SharedComponents/Cover/Cover.tsx
