+++ src/components/SharedComponents/Cover/index.ts
